--- BallaDream/src/components/common/button/button.tsx
@@ -1,7 +1,0 @@
-import * as S from '@/components/common/button/button.style';
-//버튼 예시
-function Button() {
-  return <S.StyledButton>button</S.StyledButton>;
-}
-
-export default Button;
