--- BallaDream/src/layouts/main/mainLayout.style.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-const Container = styled.div`
-  height: 100%;
-  min-height: 100vh;
-  display: flex;
-  width: 100vw;
-`;
-
-export { Container };
