--- BallaDream/src/pages/loginPage/loginPage.tsx
@@ -1,3 +1,0 @@
-export default function loginPage() {
-  return <>로그인</>;
-}
