--- BallaDream/src/pages/myPage/myPage.tsx
@@ -1,3 +1,0 @@
-export default function myPage() {
-  return <>마이</>;
-}
