--- BallaDream/src/pages/resultPage/resultPage.tsx
@@ -1,3 +1,0 @@
-export default function resultPage() {
-  return <>분석결과</>;
-}
