--- BallaDream/src/pages/uploadPage/uploadPage.style.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledButton = styled.button`
-  justify-content: center;
-  align-items: center;
-  gap: 8px;
-`;
