--- BallaDream/src/pages/uploadPage/uploadPage.tsx
@@ -1,3 +1,0 @@
-export default function UploadPage() {
-  return <>사진 업로드</>;
-}
