--- BallaDream/src/store/hooks.ts
@@ -1,7 +1,0 @@
-import type { TypedUseSelectorHook } from 'react-redux';
-import { useDispatch, useSelector } from 'react-redux';
-
-import type { TAppDispatch, TRootState } from './index';
-
-export const useAppDispatch = () => useDispatch<TAppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<TRootState> = useSelector;
