--- BallaDream/src/utils/validate.ts
@@ -1,3 +1,0 @@
-export const validate = () => {
-  return;
-};
