+++ src/components/common/button/button.tsx
@@ -1,0 +1,7 @@
+import * as S from '@/components/common/button/button.style';
+//버튼 예시
+function Button() {
+  return <S.StyledButton>button</S.StyledButton>;
+}
+
+export default Button;
