+++ src/pages/homePage/homePage.tsx
@@ -1,0 +1,9 @@
+import Icon from '@/assets/icons/img_icon.svg?react';
+
+export default function HomePage() {
+  return (
+    <>
+      홈<Icon />
+    </>
+  );
+}
