+++ src/pages/loginPage/loginPage.tsx
@@ -1,0 +1,3 @@
+export default function loginPage() {
+  return <>로그인</>;
+}
