+++ src/pages/myPage/myPage.tsx
@@ -1,0 +1,3 @@
+export default function myPage() {
+  return <>마이</>;
+}
