+++ src/pages/resultPage/resultPage.style.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const StyledButton = styled.button`
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+`;
