+++ src/pages/resultPage/resultPage.tsx
@@ -1,0 +1,3 @@
+export default function resultPage() {
+  return <>분석결과</>;
+}
