+++ src/pages/uploadPage/uploadPage.tsx
@@ -1,0 +1,3 @@
+export default function UploadPage() {
+  return <>사진 업로드</>;
+}
