+++ src/utils/validate.ts
@@ -1,0 +1,3 @@
+export const validate = () => {
+  return;
+};
